requirements.txt
gradio>=4.14.0
openai>=1.3.5
pandas>=2.1.3
numpy>=1.26.0
spacy>=3.7.2
scikit-learn>=1.3.2
python-dotenv>=1.0.0
requests>=2.31.0